+++ pkg/services/ngalert/remote/client/mimir.go
@@ -14,10 +14,11 @@ import (
 
 	alertingNotify "github.com/grafana/alerting/notify"
 
+	alertingInstrument "github.com/grafana/alerting/http/instrument"
+
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/infra/tracing"
 	apimodels "github.com/grafana/grafana/pkg/services/ngalert/api/tooling/definitions"
-	"github.com/grafana/grafana/pkg/services/ngalert/client"
 	"github.com/grafana/grafana/pkg/services/ngalert/metrics"
 	"github.com/grafana/grafana/pkg/util/httpclient"
 )
@@ -41,7 +42,7 @@ type MimirClient interface {
 }
 
 type Mimir struct {
-	client   client.Requester
+	client   alertingInstrument.Requester
 	endpoint *url.URL
 	logger   log.Logger
 	metrics  *metrics.RemoteAlertmanager
@@ -98,8 +99,8 @@ func New(cfg *Config, metrics *metrics.RemoteAlertmanager, tracer tracing.Tracer
 	c := &http.Client{
 		Transport: rt,
 	}
-	tc := client.NewTimedClient(c, metrics.RequestLatency)
-	trc := client.NewTracedClient(tc, tracer, "remote.alertmanager.client")
+	tc := alertingInstrument.NewTimedClient(c, metrics.RequestLatency)
+	trc := alertingInstrument.NewTracedClient(tc, tracer, "remote.alertmanager.client")
 
 	return &Mimir{
 		endpoint: cfg.URL,
